chore: moved events to be used for both domains
