+++ src/domain/daemon/mod.rs
@@ -1,0 +1,1 @@
+pub mod namespace;
