+++ src/domain/management/mod.rs
@@ -1,2 +1,1 @@
 pub mod project;
-pub mod events;
