+++ src/domain/mod.rs
@@ -1,1 +1,3 @@
+pub mod daemon;
 pub mod management;
+pub mod events;
